Fix presigned upload tests to use S3 client in CI
HTTP PUT to external presigned URLs fails inside Docker CI with
httpx.ReadError. Upload test fixtures via the S3 client using the
presigned key instead, and harden object_exists for NoSuchKey responses.
